moved CLT
