+++ output/CLT_0.csv
@@ -1,21 +1,21 @@
 gender,race/ethnicity,parental level of education,lunch,test preparation course,math score,reading score,writing score
-male,group C,some college,standard,completed,64,62,63
-male,group C,associate's degree,standard,none,66,68,64
-male,group A,master's degree,standard,none,65,68,65
-male,group C,master's degree,free/reduced,none,65,63,62
-female,group C,high school,free/reduced,none,51,67,65
-male,group C,associate's degree,free/reduced,none,76,67,63
-male,group B,high school,standard,none,66,55,52
-female,group D,some college,free/reduced,none,53,64,62
-male,group C,some college,free/reduced,none,61,59,59
-female,group D,some high school,free/reduced,none,46,67,66
-male,group D,some college,standard,none,35,28,28
-female,group B,master's degree,standard,none,81,97,89
-female,group D,associate's degree,standard,none,50,57,55
-female,group C,some high school,free/reduced,none,45,61,50
-female,group C,associate's degree,standard,none,65,75,79
-female,group C,high school,standard,none,83,77,80
-female,group C,some college,standard,none,53,74,59
-female,group B,some college,free/reduced,completed,65,81,74
-female,group D,some high school,free/reduced,completed,76,83,80
-male,group C,some high school,standard,none,90,90,80
+male,group D,high school,standard,none,75,69,64
+male,group A,some high school,free/reduced,completed,35,42,39
+female,group B,some college,standard,none,52,64,65
+male,group C,bachelor's degree,free/reduced,none,54,58,54
+male,group D,high school,standard,none,86,85,82
+male,group B,associate's degree,standard,completed,50,52,51
+female,group E,high school,standard,none,74,75,74
+male,group C,bachelor's degree,free/reduced,none,68,55,52
+female,group B,high school,free/reduced,completed,61,70,74
+male,group E,master's degree,free/reduced,completed,92,80,85
+female,group D,associate's degree,free/reduced,none,50,68,63
+female,group C,bachelor's degree,standard,completed,81,93,91
+female,group C,some college,standard,none,57,72,75
+female,group A,associate's degree,standard,completed,79,83,87
+female,group D,bachelor's degree,standard,none,97,100,100
+female,group E,some college,free/reduced,none,71,84,83
+female,group C,high school,standard,completed,81,81,89
+male,group E,some college,free/reduced,none,67,61,59
+female,group B,some college,standard,none,70,80,79
+male,group C,associate's degree,standard,completed,78,86,83
