+++ output/CLT_2.csv
@@ -1,21 +1,21 @@
 gender,race/ethnicity,parental level of education,lunch,test preparation course,math score,reading score,writing score
-female,group C,high school,standard,completed,45,58,59
-male,group B,high school,free/reduced,completed,75,82,80
-male,group C,high school,standard,none,82,78,73
-male,group D,bachelor's degree,free/reduced,completed,94,89,89
-female,group C,bachelor's degree,standard,none,59,64,64
-female,group B,some high school,free/reduced,none,53,64,63
-female,group C,bachelor's degree,standard,none,57,56,60
-female,group E,bachelor's degree,standard,none,77,91,83
-male,group C,some college,standard,none,85,84,79
-female,group A,some college,standard,none,71,86,80
-female,group D,associate's degree,standard,completed,55,62,68
-female,group C,associate's degree,standard,none,50,59,52
-female,group B,bachelor's degree,free/reduced,none,64,74,72
-male,group D,some college,standard,none,55,51,46
-male,group C,some high school,standard,none,59,56,54
-male,group B,some college,standard,completed,52,47,55
-female,group D,master's degree,free/reduced,none,67,75,82
-male,group D,some college,standard,none,66,66,64
-female,group B,master's degree,standard,none,81,97,89
-female,group B,high school,standard,completed,65,85,84
+female,group C,some college,standard,none,48,49,51
+male,group E,some college,free/reduced,completed,56,57,57
+female,group D,associate's degree,standard,completed,73,78,75
+male,group D,some college,standard,none,73,62,60
+female,group C,high school,free/reduced,none,51,67,65
+male,group E,bachelor's degree,standard,none,76,75,72
+male,group E,some high school,free/reduced,none,85,78,75
+male,group D,some high school,free/reduced,none,49,47,41
+male,group D,high school,standard,completed,77,73,73
+male,group D,some high school,standard,none,71,66,59
+female,group D,some high school,standard,none,78,78,83
+male,group C,associate's degree,standard,completed,89,85,84
+female,group D,high school,standard,completed,68,74,79
+female,group E,associate's degree,standard,none,82,77,84
+male,group E,some college,free/reduced,completed,83,80,76
+female,group E,high school,standard,none,82,81,73
+female,group D,master's degree,standard,none,73,78,82
+female,group C,high school,standard,none,56,53,53
+male,group D,some high school,standard,none,52,49,47
+male,group D,some college,standard,none,61,54,51
