+++ output/CLT_4.csv
@@ -1,21 +1,21 @@
 gender,race/ethnicity,parental level of education,lunch,test preparation course,math score,reading score,writing score
-female,group C,master's degree,standard,none,88,100,100
+male,group B,high school,standard,none,46,41,45
+male,group E,bachelor's degree,standard,none,77,61,59
+female,group E,associate's degree,free/reduced,none,48,52,58
+female,group D,some high school,standard,none,66,66,70
+male,group B,some college,free/reduced,completed,64,62,60
+male,group D,associate's degree,standard,completed,72,80,73
+female,group E,high school,standard,none,60,65,60
+female,group D,high school,standard,none,79,90,88
+female,group D,some college,standard,completed,52,57,63
+male,group C,some high school,standard,none,59,56,54
+male,group C,high school,standard,none,78,68,68
+male,group D,master's degree,free/reduced,completed,65,60,65
+male,group E,bachelor's degree,standard,none,76,75,72
+female,group A,some college,standard,none,71,86,80
 male,group D,high school,standard,none,74,74,70
-male,group C,some college,standard,none,76,75,68
-female,group D,some college,free/reduced,none,67,81,87
-female,group D,some high school,free/reduced,completed,52,66,74
-female,group B,bachelor's degree,free/reduced,none,54,69,68
-female,group B,associate's degree,free/reduced,none,44,56,51
-male,group B,high school,standard,none,66,55,52
-female,group C,high school,standard,none,53,58,57
-female,group E,bachelor's degree,standard,completed,100,100,100
-female,group E,some high school,standard,completed,74,70,72
-female,group B,associate's degree,standard,none,64,77,75
-male,group E,high school,standard,none,82,66,60
-female,group B,associate's degree,free/reduced,none,54,67,63
-male,group A,some high school,standard,completed,68,59,64
-female,group E,high school,free/reduced,none,50,59,53
-male,group D,some college,free/reduced,none,43,47,40
-male,group B,master's degree,standard,none,81,67,67
-female,group C,some high school,standard,completed,86,92,92
-female,group E,bachelor's degree,standard,none,77,91,83
+female,group D,high school,free/reduced,none,78,75,77
+male,group C,associate's degree,standard,none,66,68,64
+male,group E,bachelor's degree,standard,none,100,95,92
+female,group B,some high school,standard,none,64,69,69
+male,group D,some high school,standard,none,68,60,54
