+++ output/CLT_6.csv
@@ -1,21 +1,21 @@
 gender,race/ethnicity,parental level of education,lunch,test preparation course,math score,reading score,writing score
-female,group D,some high school,free/reduced,none,46,67,66
-female,group B,some college,standard,none,55,56,62
-female,group D,some college,standard,none,87,91,89
-female,group B,high school,free/reduced,none,66,74,68
-female,group E,high school,standard,none,89,99,93
-female,group E,high school,standard,none,78,78,73
-male,group B,high school,standard,none,55,41,41
-female,group E,master's degree,free/reduced,completed,88,84,89
-male,group E,bachelor's degree,standard,none,77,61,59
-male,group C,master's degree,standard,none,82,89,80
-female,group E,associate's degree,standard,completed,63,58,60
-male,group C,high school,free/reduced,none,51,52,49
-male,group C,high school,standard,completed,95,100,95
-female,group B,high school,standard,completed,68,77,82
-female,group C,high school,free/reduced,none,51,67,65
-male,group C,associate's degree,standard,none,71,73,65
-male,group B,some college,standard,none,52,53,48
-female,group D,some college,standard,none,65,63,70
-female,group C,some college,standard,completed,80,89,90
-male,group D,high school,standard,completed,54,58,56
+male,group C,bachelor's degree,free/reduced,none,68,55,52
+female,group D,some high school,standard,completed,53,64,67
+female,group D,bachelor's degree,standard,none,78,100,96
+male,group C,high school,free/reduced,completed,64,67,66
+female,group C,bachelor's degree,standard,none,92,86,94
+male,group E,associate's degree,free/reduced,completed,100,91,91
+male,group B,master's degree,standard,completed,100,100,98
+female,group C,bachelor's degree,standard,none,59,64,64
+male,group E,associate's degree,standard,none,73,62,66
+male,group D,some high school,standard,completed,69,63,60
+female,group E,bachelor's degree,standard,none,49,49,55
+male,group B,high school,standard,none,55,54,57
+male,group D,bachelor's degree,free/reduced,none,51,42,42
+female,group B,some college,standard,completed,42,64,59
+male,group E,associate's degree,free/reduced,none,59,49,49
+female,group B,some high school,standard,none,59,71,61
+male,group A,high school,standard,none,78,82,68
+male,group C,high school,standard,completed,81,72,69
+female,group D,some college,standard,none,78,79,89
+male,group C,some college,free/reduced,completed,63,54,58
