+++ output/CLT_7.csv
@@ -1,21 +1,21 @@
 gender,race/ethnicity,parental level of education,lunch,test preparation course,math score,reading score,writing score
-female,group C,high school,standard,completed,50,67,63
-male,group C,some college,standard,completed,74,76,78
-female,group B,master's degree,standard,completed,69,76,78
-male,group E,high school,free/reduced,none,86,74,78
-male,group E,bachelor's degree,standard,none,77,61,59
-male,group E,high school,free/reduced,none,61,51,47
-female,group C,high school,standard,completed,81,81,89
-female,group D,some college,standard,completed,52,57,63
-male,group C,bachelor's degree,free/reduced,none,68,55,52
-male,group C,some high school,free/reduced,none,41,44,40
-male,group E,high school,free/reduced,completed,68,54,56
+female,group D,some college,standard,none,65,63,70
+male,group C,high school,free/reduced,none,65,62,57
+male,group E,high school,standard,completed,93,82,82
+female,group C,some high school,free/reduced,none,41,57,51
+female,group C,some college,standard,completed,45,56,65
+male,group D,some high school,free/reduced,completed,49,54,56
+male,group B,some college,standard,none,52,53,48
+male,group C,bachelor's degree,standard,none,75,84,72
+female,group E,high school,standard,none,90,82,83
+male,group D,some college,free/reduced,none,82,69,73
+male,group D,high school,standard,none,83,71,68
+male,group D,some high school,standard,none,76,74,72
+male,group C,bachelor's degree,free/reduced,none,73,69,74
+male,group D,some college,standard,none,95,82,77
+male,group B,bachelor's degree,free/reduced,completed,60,68,60
+female,group C,some high school,free/reduced,none,70,77,79
 male,group D,bachelor's degree,free/reduced,completed,94,89,89
+female,group D,master's degree,free/reduced,none,67,75,82
+female,group E,some college,free/reduced,none,70,79,73
 male,group B,some high school,standard,none,78,62,59
-male,group A,high school,free/reduced,none,86,78,66
-female,group E,high school,standard,none,83,96,89
-male,group C,associate's degree,standard,none,55,60,56
-male,group D,high school,standard,completed,54,58,56
-female,group B,some college,standard,none,70,80,79
-male,group A,some high school,free/reduced,none,48,37,33
-male,group B,some college,standard,completed,100,91,90
