+++ output/CLT_8.csv
@@ -1,21 +1,21 @@
 gender,race/ethnicity,parental level of education,lunch,test preparation course,math score,reading score,writing score
-female,group A,some high school,standard,completed,64,71,76
-female,group B,some college,standard,completed,84,98,100
-female,group C,associate's degree,free/reduced,none,71,82,80
-male,group D,bachelor's degree,standard,none,76,72,79
-female,group C,high school,free/reduced,none,59,70,71
-female,group C,some college,standard,none,57,72,75
-male,group B,some high school,standard,completed,73,60,59
-female,group B,master's degree,standard,completed,69,76,78
-female,group C,associate's degree,standard,none,65,75,73
-male,group D,some high school,standard,none,52,53,50
-female,group C,some college,standard,none,84,96,96
-female,group E,some high school,free/reduced,completed,63,77,73
-female,group C,associate's degree,standard,none,38,44,44
-female,group C,high school,standard,none,58,61,62
-female,group C,some college,standard,none,62,81,69
-male,group C,high school,free/reduced,completed,46,51,45
-female,group B,some college,standard,none,51,60,59
-female,group B,some college,standard,completed,42,64,59
-male,group C,high school,free/reduced,none,50,54,49
-female,group B,high school,free/reduced,none,68,82,76
+male,group D,some college,standard,none,35,28,28
+male,group C,high school,standard,none,76,79,70
+male,group C,some college,standard,none,66,64,61
+male,group B,high school,free/reduced,none,68,75,66
+female,group C,associate's degree,free/reduced,none,68,67,68
+male,group D,high school,standard,none,75,69,64
+male,group C,some high school,standard,none,59,56,54
+male,group C,some college,standard,none,75,74,75
+female,group A,some college,standard,none,95,92,91
+female,group D,associate's degree,free/reduced,completed,74,87,87
+male,group C,associate's degree,free/reduced,none,85,88,77
+male,group C,bachelor's degree,free/reduced,completed,66,72,65
+male,group D,some college,standard,none,75,72,72
+male,group B,some high school,free/reduced,completed,58,54,53
+female,group C,high school,standard,none,63,83,75
+female,group D,some college,standard,completed,67,76,80
+female,group E,some college,standard,completed,100,100,100
+male,group B,high school,free/reduced,completed,75,82,80
+female,group C,some high school,standard,completed,68,77,80
+male,group B,some college,standard,none,62,55,49
