+++ output/IE.csv
@@ -1,21 +1,21 @@
 gender,race/ethnicity,parental level of education,lunch,test preparation course,math score,reading score,writing score
-male,group C,master's degree,standard,none,82,89,80
-female,group E,high school,standard,none,89,99,93
-female,group C,associate's degree,standard,none,71,82,78
-female,group D,high school,free/reduced,none,69,92,91
-female,group D,high school,free/reduced,none,52,63,66
-male,group D,high school,standard,none,83,71,68
-female,group C,master's degree,standard,none,74,88,78
-female,group E,some college,standard,none,53,52,53
-male,group C,associate's degree,standard,none,66,68,64
-female,group E,bachelor's degree,standard,none,95,100,97
-female,group B,associate's degree,standard,none,64,77,75
-male,group C,some college,standard,none,67,51,50
-male,group C,some high school,free/reduced,completed,32,40,43
-male,group B,some college,standard,none,71,70,68
-male,group B,some high school,standard,completed,75,72,75
-male,group E,bachelor's degree,free/reduced,none,49,48,46
-female,group D,associate's degree,standard,completed,47,60,60
-female,group C,some college,standard,none,89,90,91
-male,group D,some high school,standard,none,78,84,81
-male,group D,some high school,free/reduced,completed,40,55,45
+female,group C,some high school,free/reduced,none,41,57,51
+male,group D,some college,free/reduced,none,52,58,51
+female,group D,high school,free/reduced,completed,80,85,89
+female,group E,some high school,standard,none,94,97,95
+female,group B,associate's degree,free/reduced,none,44,56,51
+male,group B,high school,standard,completed,85,85,85
+male,group C,associate's degree,standard,completed,86,83,77
+male,group D,some high school,standard,completed,69,63,60
+female,group D,high school,standard,none,54,67,59
+male,group B,bachelor's degree,standard,none,84,77,79
+male,group C,associate's degree,standard,none,55,60,56
+female,group A,associate's degree,standard,completed,67,78,81
+male,group C,some college,standard,none,85,84,79
+male,group C,some high school,standard,completed,64,63,63
+male,group D,high school,standard,none,85,83,79
+male,group B,bachelor's degree,standard,none,65,62,59
+male,group C,some college,standard,completed,72,81,80
+female,group E,some college,standard,none,67,70,67
+female,group D,some high school,free/reduced,none,46,67,66
+female,group A,master's degree,standard,completed,74,91,92
